* android-12.0/base: (29 commits)
  Makefile: Fix GCC_TOOLCHAIN_DIR prefix for Clang cross compilation
  Makefile: Silence spam
  OplusKernelEnvConfig: Don't print allowed macros as warnings
  ARM64/configs: RM6785: Set localversion to `lineage`
  ARM64/configs: RM6785: Enable backlight trigger
  drivers: mtk-connectivity: Globally silence connectivity logspam
  drivers: misc-mtk: Import MTK modules from begonia-r-oss
  ARM64/configs: RM6785: Initial defconfig
  ARM64/dts: mediatek: Fix dtbo building
  DrvGen: Force use of py2
  [SQUASH] treewide: Add back drvgen and dtbo building
  treewide: Stop using kernel-4.14
  treewide: Import oplus changes
  sound: Import oplus changes
  security: Import oplus changes
  net: Import oplus changes
  mm: Import oplus changes
  lib: Import oplus changes
  kernel: Import oplus changes
  init: Import oplus changes
  ...

+++ drivers/misc/mediatek/pmic/mt6359p/v1/pmic_throttling_dlpt.c
@@ -413,6 +413,9 @@ static irqreturn_t fg_cur_h_int_handler(int irq, void *data)
 		, pmic_get_register_value(PMIC_RG_INT_EN_FG_CUR_H)
 		, pmic_get_register_value(PMIC_RG_INT_EN_FG_CUR_L));
 #else
+//#ifdef VENDOR_EDIT
+/*Dongru.Zhao@BSP.CHG.Basic, 2020/07/27, zdr Add for kernel log .*/
+/*
 	PMICLOG("Reg[0x%x]=0x%x, Reg[0x%x]=0x%x, Reg[0x%x]=0x%x\n",
 		PMIC_FG_CUR_HTH_ADDR,
 		upmu_get_reg_value(PMIC_FG_CUR_HTH_ADDR),
@@ -420,6 +423,8 @@ static irqreturn_t fg_cur_h_int_handler(int irq, void *data)
 		upmu_get_reg_value(PMIC_FG_CUR_LTH_ADDR),
 		PMIC_RG_INT_EN_FG_CUR_H_ADDR,
 		upmu_get_reg_value(PMIC_RG_INT_EN_FG_CUR_H_ADDR));
+*/
+//#endif
 #endif
 	return IRQ_HANDLED;
 }
@@ -444,6 +449,9 @@ static irqreturn_t fg_cur_l_int_handler(int irq, void *data)
 		, pmic_get_register_value(PMIC_RG_INT_EN_FG_CUR_H)
 		, pmic_get_register_value(PMIC_RG_INT_EN_FG_CUR_L));
 #else
+//#ifdef VENDOR_EDIT
+/*Dongru.Zhao@BSP.CHG.Basic, 2020/07/27, zdr Add for kernel log .*/
+/*
 	PMICLOG("Reg[0x%x]=0x%x, Reg[0x%x]=0x%x, Reg[0x%x]=0x%x\n",
 		PMIC_FG_CUR_HTH_ADDR,
 		upmu_get_reg_value(PMIC_FG_CUR_HTH_ADDR),
@@ -451,6 +459,8 @@ static irqreturn_t fg_cur_l_int_handler(int irq, void *data)
 		upmu_get_reg_value(PMIC_FG_CUR_LTH_ADDR),
 		PMIC_RG_INT_EN_FG_CUR_H_ADDR,
 		upmu_get_reg_value(PMIC_RG_INT_EN_FG_CUR_H_ADDR));
+*/
+//#endif
 #endif
 	return IRQ_HANDLED;
 }
@@ -619,7 +629,7 @@ void register_battery_percent_notify_ext(
 {
 	PMICLOG("[%s] start\n", __func__);
 
-	bpcb_tb_ext[(unsigned int)prio_val].bpcb = battery_percent_callback;
+	bpcb_tb_ext[prio_val].bpcb = battery_percent_callback;
 
 	pr_info("[%s] prio_val=%d\n", __func__, prio_val);
 
@@ -1879,7 +1889,9 @@ void pmic_throttling_dlpt_suspend(void)
 #ifdef BATTERY_OC_PROTECT
 	disable_irq_nosync(fg_cur_h_irq);
 	disable_irq_nosync(fg_cur_l_irq);
-
+//#ifdef VENDOR_EDIT
+/*Dongru.Zhao@BSP.CHG.Basic, 2020/07/27, zdr Add for kernel log .*/
+/*
 	PMICLOG("Reg[0x%x]=0x%x, Reg[0x%x]=0x%x, Reg[0x%x]=0x%x\n",
 		PMIC_FG_CUR_HTH_ADDR,
 		upmu_get_reg_value(PMIC_FG_CUR_HTH_ADDR),
@@ -1887,6 +1899,8 @@ void pmic_throttling_dlpt_suspend(void)
 		upmu_get_reg_value(PMIC_FG_CUR_LTH_ADDR),
 		PMIC_RG_INT_EN_FG_CUR_H_ADDR,
 		upmu_get_reg_value(PMIC_RG_INT_EN_FG_CUR_H_ADDR));
+*/
+//#endif
 #endif
 }
 
@@ -1896,7 +1910,9 @@ void pmic_throttling_dlpt_resume(void)
 #ifdef BATTERY_OC_PROTECT
 	enable_irq(fg_cur_h_irq);
 	enable_irq(fg_cur_l_irq);
-
+//#ifdef VENDOR_EDIT
+/*Dongru.Zhao@BSP.CHG.Basic, 2020/07/27, zdr Add for kernel log .*/
+/*
 	PMICLOG("Reg[0x%x]=0x%x, Reg[0x%x]=0x%x, Reg[0x%x]=0x%x\n",
 		PMIC_FG_CUR_HTH_ADDR,
 		upmu_get_reg_value(PMIC_FG_CUR_HTH_ADDR),
@@ -1904,6 +1920,8 @@ void pmic_throttling_dlpt_resume(void)
 		upmu_get_reg_value(PMIC_FG_CUR_LTH_ADDR),
 		PMIC_RG_INT_EN_FG_CUR_H_ADDR,
 		upmu_get_reg_value(PMIC_RG_INT_EN_FG_CUR_H_ADDR));
+*/
+//#endif
 #endif
 }
